+++ gamma_dashboard/dashboard/api/v0/urls.py
@@ -5,6 +5,7 @@ from django.urls import re_path
 
 from gamma_dashboard.constants import COURSE_ID_PATTERN
 from gamma_dashboard.dashboard.api.v0.views import (
+    BadgeLeaderboardApiView,
     GameProfileApiView,
     GameUserAvatarConfigApiView,
     LeaderboardApiView,
@@ -13,6 +14,11 @@ from gamma_dashboard.dashboard.api.v0.views import (
 
 
 urlpatterns = [
+    re_path(
+        fr'^leaderboard/badge/(?P<badge_slug>[-\w]+)(?:/{COURSE_ID_PATTERN})?/?$',
+        BadgeLeaderboardApiView.as_view(),
+        name='api-badge-leaderboard',
+    ),
     re_path(fr'^leaderboard(?:/{COURSE_ID_PATTERN})?/?$', LeaderboardApiView.as_view(), name='api-leaderboard'),
     re_path(r'^game-profile/', GameProfileApiView.as_view(), name='api-gameprofile'),
     re_path(r'^user-badges/(?P<username>[\w.@+-]+)/?$', UserBadgesApiView.as_view(), name='api-user-badges'),