+++ gamma_dashboard/static/dashboard/js/app/src/dashboard/components/progress-badge/ProgressBadgesModal.jsx
@@ -43,6 +43,7 @@ const ProgressBadgesModal = ({
           {filteredActiveBadges.map((item) => (
             <ProgressBadge
               key={getItemDataFunction(item).title}
+              slug={item[0]}
               data={getItemDataFunction(item)}
               center
             />