+++ gamma_dashboard/static/dashboard/js/app/src/leaderboard/index.jsx
@@ -1,1 +1,2 @@
 export { default as LeaderboardPage } from './LeaderboardPage';
+export { default as BadgeLeaderboardPage } from './BadgeLeaderboardPage';